requirements.txt
openai
python-dotenv
requests
googlesearch-python
lxml
fastapi
uvicorn